+++ packages/stupid_simple_sheet/CHANGELOG.md
@@ -1,3 +1,32 @@
+## 0.5.0
+
+> Note: This release has breaking changes.
+
+ - **BREAKING** **FIX**: don't call `Navigator.did[Start|Stop]UserGesture` by default to match Flutter sheets.
+
+    You can restore the old behavior by passing `callNavigatorUserGestureMethods = true` to your sheet route
+
+
+## 0.4.2
+
+ - **FEAT**: add `StupidSimpleSheetController` that can be used from a sheets subtree for imperative control.
+
+## 0.4.1
+
+ - **FIX**: velocity scaling when overdragging with resistance.
+ - **FIX**: secondary animation in cupertino sheet.
+ - **FEAT**: cupertino sheet can now be dragged over its limits with resistance.
+ - **FEAT**: sheets support snapping points now.
+ - **FEAT**: cupertino sheet top padding is based on safe area now.
+
+## 0.4.0+2
+
+ - Update a dependency to the latest release.
+
+## 0.4.0+1
+
+ - Update a dependency to the latest release.
+
 ## 0.4.0
 
 > Note: This release has breaking changes.